Experimental support for linking all exts as static libs...
...into the main DLL.

Note: To compile the extensions as static libs use

--with-static-linked-ext with win32\configure.bat

The extensions need to be initialized and provided as "Features" (via
rb_provide) so that require 'Xyz' doesn't look for their files (as they no
longer exist, but are builtin).

Needed to stub out the initialization for miniruby as miniruby is built
during the build process and is used to generate the makefiles for the
extensions: This means miniruby cannot depend on the extensions (it would
be a catch 22). Miniruby thus loads no extensions (which was always the
case anyway, by definition).

Another subtlety is the need to initialize the "enc" library before all
other extensions -- this was previously done by the interpreter itself in
process_options(). But that happens well after ruby_setup(). Since the
extensions themselves depend on the encoding library we need to do it in
rb_call_inits_ext, and *before* all other inits in this function.

One more thing: Since the extensions themselves do thigns like rb_require
internally, they require the load_path to be established. Previously,
extensions would only be loaded well after ruby_setup had completed, so
the load_path would be all setup correctly. But now the extensions are
loaded *during* ruby_setup, and *before* the load_path is properly
established.

This means that we need to establish the path ourselves prior to
initializing the modules.

Currently we set it to the 3 entries that are normally passed as part of
the build when installing (nmake install). E.g.

.\ruby.exe -I./lib -I".ext/x64-mswin64_120" -I. "./tool/runruby.rb"
--extout=".ext"  -- --disable-gems -r./x64-mswin64_120-fake
./tool/rbinstall.rb --make="nmake" --dest-dir=""  --extout=".ext"
--mflags="-l"  --make-flags="                  "  --data-mode=0644
--prog-mode=0755  --installed-list .installed.list  --mantype=""

Note the three -I switches after ruby.exe: the lib directory, the
.ext/<arch> directory and the current directory (-I.).

We mimick that in rb_call_inits_ext to allow nmake install to succeed, and
to let ruby.exe run from its built location.

Likely we'll have to add more to this list to support running it from its
usual installed location (i.e. see ruby_initial_load_paths)