Add support for locking down core extensions.
This change introduces a new API, ruby_set_ext_filter, that embedders can use to
register a callback that will filter the extensions that are added to the
VM's loading_table. This means those extensions will not be found with 'require'
et al.

This can be used, for example, to lock down the socket.so extension.

By default there is no filter, so nothing is locked down.